make secure-password optional peer dependency
