+++ packages/blitz-auth/package.json
@@ -34,12 +34,17 @@
     "nanoid": "3.2.0",
     "passport": "0.5.2",
     "path": "0.12.7",
-    "secure-password": "4.0.0",
     "supports-color": "8.1.1",
     "url": "0.11.0"
   },
   "peerDependencies": {
-    "blitz": "2.0.0-beta.22"
+    "blitz": "2.0.0-beta.22",
+    "secure-password": "4.0.0"
+  },
+  "peerDependenciesMeta": {
+    "secure-password": {
+      "optional": true
+    }
   },
   "devDependencies": {
     "@blitzjs/config": "workspace:2.0.0-beta.22",
@@ -53,6 +58,7 @@
     "blitz": "2.0.0-beta.22",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "secure-password": "4.0.0",
     "typescript": "^4.8.4",
     "unbuild": "0.7.6",
     "watch": "1.0.2"
